+++ packages/hoc-input/src/formInput.story.js
@@ -70,7 +70,12 @@ storiesOf("HOC formInput", "module")
     "html disabled input with label",
     withInfo()(() => (
       <div>
-        <EnhancedInput disabled label="Input Label" type="text" />
+        <EnhancedInput
+          disabled
+          label="Input Label"
+          type="text"
+          value="Disabled field"
+        />
       </div>
     ))
   )
@@ -78,4 +83,12 @@ storiesOf("HOC formInput", "module")
   .add(
     "html search input with value",
     withInfo()(() => <EnhancedInput type="search" value="broccoli" />)
+  )
+  .add(
+    "html text input with combobox role",
+    withInfo()(() => (
+      <div style={{ width: "300px" }}>
+        <EnhancedInput role="combobox" placeholder="Select value" />
+      </div>
+    ))
   );