+++ packages/dropdown/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-dropdown@0.4.1...@crave/farmblocks-dropdown@0.5.0) (2018-02-07)
+
+
+### Features
+
+* **dropdown:** export dropdown wrappers ([0469a73](https://github.com/CraveFood/farmblocks/commit/0469a73))
+* **dropdown item:** add selected and highlighted styles to dropdonw item wrapper ([3add0cb](https://github.com/CraveFood/farmblocks/commit/3add0cb))
+
+
+
+
 <a name="0.4.1"></a>
 ## [0.4.1](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-dropdown@0.4.0...@crave/farmblocks-dropdown@0.4.1) (2018-01-29)